Fix warning

+++ src/kshim.cpp
@@ -144,7 +144,7 @@ string KShim::binaryName()
 int KShim::run(const KShimData &data, int argc, char *argv[])
 {
     std::vector<std::string> args;
-    args.reserve(argc);
+    args.reserve(static_cast<size_t>(argc));
     for (int i = 1; i < argc; ++i)
     {
         args.push_back(argv[i]);
